modernize JS_pp200_signal for run15

+++ submit/JS_pp200_signal/pass3_mbdepd/condor/restart_held.sh
@@ -2,16 +2,27 @@
 if [ $# -eq 0 ]
 then
   echo "No arguments supplied"
-  exit 1
+  exit 0
 fi
+run=15
+runnumber=$(printf "%010d" $run)
+MHz=2
 echo $1
 variable=$1
-#exit 0
-condor_q | grep ' H ' | grep run_pass3_bbcepd_js.sh | grep ${variable}  > bla
+
+
+condor_q | grep ' H ' | grep run_pass3_mbdepd_js.sh | grep ${variable} > bla
+
+[ -s bla ] ||  exit 1
+
 for i in `cat bla| awk '{print $1}'`; do condor_rm $i; done
-if [ -f tmplist ]
-then
-rm tmplist
-fi
-for i in `cat bla | awk '{print $12}' | awk -F- '{print $3}' | awk -F. '{print "-0000000008-"$1".job"}'`; do echo $i >> tmplist ; done
-for i in `cat tmplist`; do condor_submit log/${variable}/condor_${variable}$i; done
+
+[ -f tmplist ] && rm tmplist
+
+for i in `cat bla | awk '{print $12}' | awk -F- '{print $3}' | awk -F. -v runnumber=${runnumber} '{print ""runnumber"-"$1".job"}'`; do echo $i >> tmplist ; done
+
+[ -f sedlist ] && rm sedlist
+
+for i in `cat tmplist`; do echo log/run${run}/${variable}_${MHz}MHz/condor_${variable}_${MHz}MHz-$i >> sedlist; done
+
+for i in `cat sedlist`; do condor_submit $i; done
